requirements.txt
# langchain
langchain
langchain_core
langchain_huggingface
langchain_community

# others
python-dotenv
sentence-transformers
faiss-cpu
streamlit